CLOUD-24: Don't remove PVCs on update
