+++ pkg/stub/replset.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/Percona-Lab/percona-server-mongodb-operator/pkg/apis/psmdb/v1alpha1"
 
-	"github.com/operator-framework/operator-sdk/pkg/sdk"
 	motPkg "github.com/percona/mongodb-orchestration-tools/pkg"
 	podk8s "github.com/percona/mongodb-orchestration-tools/pkg/pod/k8s"
 	"github.com/sirupsen/logrus"
@@ -41,21 +40,6 @@ func getReplsetDialInfo(m *v1alpha1.PerconaServerMongoDB, replset *v1alpha1.Repl
 	}
 }
 
-// getReplsetStatus returns a ReplsetStatus object for a given replica set
-func getReplsetStatus(m *v1alpha1.PerconaServerMongoDB, replset *v1alpha1.ReplsetSpec) *v1alpha1.ReplsetStatus {
-	for _, rs := range m.Status.Replsets {
-		if rs.Name == replset.Name {
-			return rs
-		}
-	}
-	status := &v1alpha1.ReplsetStatus{
-		Name:    replset.Name,
-		Members: []*v1alpha1.ReplsetMemberStatus{},
-	}
-	m.Status.Replsets = append(m.Status.Replsets, status)
-	return status
-}
-
 // isReplsetInitialized returns a boolean reflecting if the replica set has been initiated
 func isReplsetInitialized(m *v1alpha1.PerconaServerMongoDB, replset *v1alpha1.ReplsetSpec, status *v1alpha1.ReplsetStatus, podList *corev1.PodList, usersSecret *corev1.Secret) bool {
 	if status.Initialized {
@@ -80,36 +64,6 @@ func isReplsetInitialized(m *v1alpha1.PerconaServerMongoDB, replset *v1alpha1.Re
 	status.Initialized = true
 
 	return true
-}
-
-// getReplsetMemberStatuses returns a list of ReplsetMemberStatus structs for a given replset
-func getReplsetMemberStatuses(m *v1alpha1.PerconaServerMongoDB, replset *v1alpha1.ReplsetSpec, pods []corev1.Pod, usersSecret *corev1.Secret) []*v1alpha1.ReplsetMemberStatus {
-	members := []*v1alpha1.ReplsetMemberStatus{}
-	for _, pod := range pods {
-		dialInfo := getReplsetDialInfo(m, replset, []corev1.Pod{pod}, usersSecret)
-		dialInfo.Direct = true
-		session, err := mgo.DialWithInfo(dialInfo)
-		if err != nil {
-			logrus.Debugf("Cannot connect to mongodb host %s: %v", dialInfo.Addrs[0], err)
-			continue
-		}
-		defer session.Close()
-		session.SetMode(mgo.Eventual, true)
-
-		logrus.Debugf("Updating status for host: %s", dialInfo.Addrs[0])
-
-		buildInfo, err := session.BuildInfo()
-		if err != nil {
-			logrus.Debugf("Cannot get buildInfo from mongodb host %s: %v", dialInfo.Addrs[0], err)
-			continue
-		}
-
-		members = append(members, &v1alpha1.ReplsetMemberStatus{
-			Name:    dialInfo.Addrs[0],
-			Version: buildInfo.Version,
-		})
-	}
-	return members
 }
 
 // handleReplsetInit runs the k8s-mongodb-initiator from within the first running pod's mongod container.
@@ -215,57 +169,15 @@ func (h *Handler) handleStatefulSetUpdate(m *v1alpha1.PerconaServerMongoDB, set 
 	return nil
 }
 
-// updateStatus updates the PerconaServerMongoDB status
-func (h *Handler) updateStatus(m *v1alpha1.PerconaServerMongoDB, replset *v1alpha1.ReplsetSpec, usersSecret *corev1.Secret) (*corev1.PodList, error) {
-	var doUpdate bool
-
-	podList := podList()
-	err := h.client.List(m.Namespace, podList, sdk.WithListOptions(getLabelSelectorListOpts(m, replset)))
-	if err != nil {
-		return nil, fmt.Errorf("failed to list pods for replset %s: %v", replset.Name, err)
-	}
-
-	// Update status pods list
-	podNames := getPodNames(podList.Items)
-	status := getReplsetStatus(m, replset)
-	if !reflect.DeepEqual(podNames, status.Pods) {
-		status.Pods = podNames
-		doUpdate = true
-	}
-
-	// Update mongodb replset member status list
-	members := getReplsetMemberStatuses(m, replset, podList.Items, usersSecret)
-	if !reflect.DeepEqual(members, status.Members) {
-		status.Members = members
-		doUpdate = true
-	}
-
-	// Send update to SDK if something changed
-	if doUpdate {
-		err = h.client.Update(m)
-		if err != nil {
-			return nil, fmt.Errorf("failed to update status for replset %s: %v", replset.Name, err)
-		}
-	}
-
-	// Update the pods list that is read by the watchdog
-	if h.pods == nil {
-		h.pods = podk8s.NewPods(m.Name, m.Namespace)
-	}
-	h.pods.SetPods(podList.Items)
-
-	return podList, nil
-}
-
 // ensureReplsetStatefulSet ensures a StatefulSet exists
-func (h *Handler) ensureReplsetStatefulSet(m *v1alpha1.PerconaServerMongoDB, replset *v1alpha1.ReplsetSpec) error {
+func (h *Handler) ensureReplsetStatefulSet(m *v1alpha1.PerconaServerMongoDB, replset *v1alpha1.ReplsetSpec) (*appsv1.StatefulSet, error) {
 	limits, err := parseSpecResourceRequirements(replset.Limits)
 	if err != nil {
-		return err
+		return nil, err
 	}
 	requests, err := parseSpecResourceRequirements(replset.Requests)
 	if err != nil {
-		return err
+		return nil, err
 	}
 	resources := corev1.ResourceRequirements{
 		Limits:   limits,
@@ -285,12 +197,12 @@ func (h *Handler) ensureReplsetStatefulSet(m *v1alpha1.PerconaServerMongoDB, rep
 
 	set, err := h.newPSMDBStatefulSet(m, replset, &resources)
 	if err != nil {
-		return err
+		return nil, err
 	}
 	err = h.client.Create(set)
 	if err != nil {
 		if !k8serrors.IsAlreadyExists(err) {
-			return err
+			return nil, err
 		}
 	} else {
 		logrus.WithFields(lf).Infof("created stateful set for replset: %s", replset.Name)
@@ -299,42 +211,52 @@ func (h *Handler) ensureReplsetStatefulSet(m *v1alpha1.PerconaServerMongoDB, rep
 	// Ensure the spec is up to date
 	err = h.handleStatefulSetUpdate(m, set, replset, &resources)
 	if err != nil {
-		return fmt.Errorf("failed to get stateful set for replset %s: %v", replset.Name, err)
+		return nil, fmt.Errorf("failed to get stateful set for replset %s: %v", replset.Name, err)
 	}
 
-	return nil
+	return set, nil
 }
 
 // ensureReplset ensures resources for a PSMDB replset exist
-func (h *Handler) ensureReplset(m *v1alpha1.PerconaServerMongoDB, podList *corev1.PodList, replset *v1alpha1.ReplsetSpec, usersSecret *corev1.Secret) (*v1alpha1.ReplsetStatus, error) {
+func (h *Handler) ensureReplset(m *v1alpha1.PerconaServerMongoDB, podList *corev1.PodList, replset *v1alpha1.ReplsetSpec, usersSecret *corev1.Secret) error {
+	status := getReplsetStatus(m, replset)
+
 	// Create the StatefulSet if it doesn't exist
-	err := h.ensureReplsetStatefulSet(m, replset)
+	set, err := h.ensureReplsetStatefulSet(m, replset)
 	if err != nil {
 		logrus.Errorf("failed to create stateful set for replset %s: %v", replset.Name, err)
-		return nil, err
+		return err
 	}
 
 	// Initiate the replset if it hasn't already been initiated + there are pods +
 	// we have waited the ReplsetInitWait period since starting
-	status := getReplsetStatus(m, replset)
 	if !isReplsetInitialized(m, replset, status, podList, usersSecret) && len(podList.Items) >= 1 && time.Since(h.startedAt) > ReplsetInitWait {
 		err = h.handleReplsetInit(m, replset, podList.Items)
 		if err != nil {
-			return nil, err
+			return err
 		}
 
 		// update status after replset init
 		status.Initialized = true
 		err = h.client.Update(m)
 		if err != nil {
-			return nil, fmt.Errorf("failed to update status for replset %s: %v", replset.Name, err)
+			return fmt.Errorf("failed to update status for replset %s: %v", replset.Name, err)
 		}
 		logrus.Infof("changed state to initialised for replset %s", replset.Name)
 
 		// ensure the watchdog is started
 		err = h.ensureWatchdog(m, usersSecret)
 		if err != nil {
-			return nil, fmt.Errorf("failed to start watchdog: %v", err)
+			return fmt.Errorf("failed to start watchdog: %v", err)
+		}
+	}
+
+	// Remove PVCs left-behind from scaling down if no update is running
+	if !isStatefulSetUpdating(set) {
+		err = h.persistentVolumeClaimReaper(m, podList.Items, replset, status)
+		if err != nil {
+			logrus.Errorf("failed to run persistent volume claim reaper for replset %s: %v", replset.Name, err)
+			return err
 		}
 	}
 
@@ -344,11 +266,11 @@ func (h *Handler) ensureReplset(m *v1alpha1.PerconaServerMongoDB, podList *corev
 	if err != nil {
 		if !k8serrors.IsAlreadyExists(err) {
 			logrus.Errorf("failed to create psmdb service: %v", err)
-			return nil, err
+			return err
 		}
 	} else {
 		logrus.Infof("created service %s", service.Name)
 	}
 
-	return status, nil
+	return nil
 }
