+++ pkg/stub/status_test.go
@@ -1,0 +1,11 @@
+package stub
+
+import (
+//	"testing"
+
+//	"github.com/stretchr/testify/assert"
+)
+
+//func TestGetReplsetStatus(t *testing.T) {}
+//func TestGetReplsetMemberStatuses(t *testing.T) {}
+//func TestHandlerUpdateStatus(t *testing.T) {}
